+++ source/deprecated/omni.isaac.dynamic_control/python/tests/test_articulation_other.py
@@ -81,7 +81,7 @@ class TestArticulationOther(omni.kit.test.AsyncTestCase):
     async def test_articulation_carter(self, gpu=False):
 
         (result, error) = await open_stage_async(
-            self._assets_root_path + "/Isaac/Robots/Carter/carter_v1_physx_lidar.usd"
+            self._assets_root_path + "/Isaac/Robots/NVIDIA/Carter/carter_v1_physx_lidar.usd"
         )
         # Make sure the stage loaded
         self.assertTrue(result)