+++ source/extensions/isaacsim.test.collection/isaacsim/test/collection/tests/test_drive_goal_carter_v2.py
@@ -41,7 +41,7 @@ class TestDriveGoalCarterv2(omni.kit.test.AsyncTestCase):
         self._extension_path = get_extension_path_from_name("isaacsim.test.collection")
 
         # add in carter (from nucleus)
-        self.usd_path = self._assets_root_path + "/Isaac/Robots/NVIDIA/Carter/nova_carter/nova_carter.usd"
+        self.usd_path = self._assets_root_path + "/Isaac/Robots/NVIDIA/NovaCarter/nova_carter.usd"
         (result, error) = await open_stage_async(self.usd_path)
         PhysicsSchemaTools.addGroundPlane(
             omni.usd.get_context().get_stage(), "/groundPlane", "Z", 1500, Gf.Vec3f(0, 0, 0), Gf.Vec3f(0.5)