+++ source/extensions/isaacsim.examples.interactive/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [4.4.15] - 2025-04-09
+### Changed
+- Update Isaac Sim NVIDIA robot asset path
+
 ## [4.4.14] - 2025-04-04
 ### Changed
 - Version bump to fix extension publishing issues