+++ source/extensions/isaacsim.sensors.physics/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [0.3.15] - 2025-04-09
+### Changed
+- Update Isaac Sim NVIDIA robot asset path
+
 ## [0.3.14] - 2025-04-04
 ### Changed
 - Version bump to fix extension publishing issues